adapt to new zvtm 0.11.0 double-prec float coord sys
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3586 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zuist-engine/trunk/src/main/java/fr/inria/zuist/engine/ClosedShapeDescription.java
@@ -115,12 +115,12 @@ public class ClosedShapeDescription extends ObjectDescription {
     }
     
     /** Get x-coordinate of object in virtual space. */
-    public long getX(){
+    public double getX(){
         return glyph.vx;
     }
     
     /** Get y-coordinate of object in virtual space. */
-    public long getY(){
+    public double getY(){
         return glyph.vy;
     }
